+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to 'Kwai\JSONAPI' will be documented here.
 
 # 1.0.1
 - Add id argument to JSONAPI\Resource attribute to allow other properties or a method to return the id.
+- An array can be passed to setMeta
 
 # 1.0.0
 First release